authorizer/requirements.txt
# No external dependencies required for now 

# Dependencies for JWT validation with AWS Cognito
pyjwt==2.8.0
cryptography>=3.4.8
requests>=2.28.0 